SITL: convert to double precision for positions

+++ libraries/SITL/SIM_Gazebo.cpp
@@ -117,9 +117,9 @@ void Gazebo::recv_fdm(const struct sitl_input &input)
                            static_cast<float>(pkt.velocity_xyz[1]),
                            static_cast<float>(pkt.velocity_xyz[2]));
 
-    position = Vector3f(static_cast<float>(pkt.position_xyz[0]),
-                        static_cast<float>(pkt.position_xyz[1]),
-                        static_cast<float>(pkt.position_xyz[2]));
+    position = Vector3d(pkt.position_xyz[0],
+                        pkt.position_xyz[1],
+                        pkt.position_xyz[2]);
 
 
     // auto-adjust to simulation frame rate
